Refactored account to use base view & presenter
